+++ handler/get_test.go
@@ -1,4 +1,4 @@
-package main
+package handler
 
 import (
 	"bytes"
@@ -9,51 +9,63 @@ import (
 	"testing"
 )
 
-func TestPostHandler_Success(t *testing.T) {
-	ts := httptest.NewServer(http.HandlerFunc(postHandler))
+func TestGetHandler_Success(t *testing.T) {
+	ts := httptest.NewServer(http.HandlerFunc(GetHandler))
 	defer ts.Close()
 
 	bodyEcho := bytes.NewBufferString("Hello world!")
-	respPost, err := http.Post(ts.URL, "application/json", bodyEcho)
-	if err != nil {
-		t.Fatal(err)
-	}
-	outPost, err := io.ReadAll(respPost.Body)
-	if err != nil {
-		t.Fatal(err)
-	}
-	stringOut := string(outPost)
 
+	respGet, err := http.NewRequest("GET", ts.URL, bodyEcho)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	respGet.Header.Set("Content-Type", "text/html")
+	client := &http.Client{}
+	req, err := client.Do(respGet)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	outGet, err := io.ReadAll(req.Body)
+	if err != nil {
+		t.Fatal(err)
+	}
+	stringOut := string(outGet)
 	if stringOut != "Hello world!" {
 		t.Fatalf(`Output should be "Hello world!",  but you have %s`, stringOut)
 	}
 }
-func TestPostHandler_Wrong(t *testing.T) {
-	ts := httptest.NewServer(http.HandlerFunc(postHandler))
+
+func TestGetHandler_Wrong(t *testing.T) {
+	ts := httptest.NewServer(http.HandlerFunc(GetHandler))
 	defer ts.Close()
 
 	bodyEcho := bytes.NewBufferString("Hello world!")
-	reqPost, err := http.NewRequest("GET", ts.URL, bodyEcho)
+
+	respPost, err := http.NewRequest(http.MethodPost, ts.URL, bodyEcho)
 	if err != nil {
 		t.Fatal(err)
 	}
+
+	respPost.Header.Set("Content-Type", "text/html")
 	client := &http.Client{}
-	req, err := client.Do(reqPost)
+	req, err := client.Do(respPost)
 	if err != nil {
 		t.Fatal(err)
 	}
+
 	outPost, err := io.ReadAll(req.Body)
 	if err != nil {
 		t.Fatal(err)
 	}
 
-	var errPost errPost
+	var errPost errResult
 	err = json.Unmarshal(outPost, &errPost)
 	if err != nil {
 		t.Fatal(err)
 	}
-
-	MsgErr := "Supports only POST method. Please use POST method."
+	MsgErr := "Supported only GET method. Please use GET method."
 	if errPost.ErrMsg != MsgErr {
 		t.Fatalf("Error should be: %s\n but i recieved: %s\n", MsgErr, errPost.ErrMsg)
 	}
