+++ handler/go.mod
@@ -1,0 +1,3 @@
+module handler
+
+go 1.19
