+++ handler/ip_test.go
@@ -1,4 +1,4 @@
-package main
+package handler
 
 import (
 	"encoding/json"
@@ -8,15 +8,15 @@ import (
 	"testing"
 )
 
-func TestUserAgent(t *testing.T) {
-	ts := httptest.NewServer(http.HandlerFunc(userAgentHandler))
+func TestIpHandler(t *testing.T) {
+	ts := httptest.NewServer(http.HandlerFunc(IpHandler))
 	defer ts.Close()
 
 	req, err := http.NewRequest("GET", ts.URL, nil)
 	if err != nil {
 		t.Fatal(err)
 	}
-	req.Header.Set("User-Agent", "Golang")
+	req.Host = "78.175.231.108"
 
 	client := &http.Client{}
 	resp, err := client.Do(req)
@@ -29,13 +29,13 @@ func TestUserAgent(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	var userJSON userResponse
-	err = json.Unmarshal(out, &userJSON)
+	var ipJSON ip
+	err = json.Unmarshal(out, &ipJSON)
 	if err != nil {
 		t.Fatal(err)
 	}
 
-	if userJSON.UserAgent != "Golang" {
-		t.Fatalf("user agent result is wrong\n\nwant: %s\n got: %s\n", "Golang", userJSON.UserAgent)
+	if ipJSON.Origin != "78.175.231.108" {
+		t.Fatalf("ip result is wrong\n\nwant: %s\n got: %s\n", "78.175.231.108", ipJSON.Origin)
 	}
 }
