+++ handler/post_test.go
@@ -1,4 +1,4 @@
-package main
+package handler
 
 import (
 	"bytes"
@@ -9,63 +9,51 @@ import (
 	"testing"
 )
 
-func TestGetHandler_Success(t *testing.T) {
-	ts := httptest.NewServer(http.HandlerFunc(getHandler))
+func TestPostHandler_Success(t *testing.T) {
+	ts := httptest.NewServer(http.HandlerFunc(PostHandler))
 	defer ts.Close()
 
 	bodyEcho := bytes.NewBufferString("Hello world!")
-
-	respGet, err := http.NewRequest("GET", ts.URL, bodyEcho)
+	respPost, err := http.Post(ts.URL, "application/json", bodyEcho)
 	if err != nil {
 		t.Fatal(err)
 	}
-
-	respGet.Header.Set("Content-Type", "text/html")
-	client := &http.Client{}
-	req, err := client.Do(respGet)
+	outPost, err := io.ReadAll(respPost.Body)
 	if err != nil {
 		t.Fatal(err)
 	}
+	stringOut := string(outPost)
 
-	outGet, err := io.ReadAll(req.Body)
-	if err != nil {
-		t.Fatal(err)
-	}
-	stringOut := string(outGet)
 	if stringOut != "Hello world!" {
 		t.Fatalf(`Output should be "Hello world!",  but you have %s`, stringOut)
 	}
 }
-
-func TestGetHandler_Wrong(t *testing.T) {
-	ts := httptest.NewServer(http.HandlerFunc(getHandler))
+func TestPostHandler_Wrong(t *testing.T) {
+	ts := httptest.NewServer(http.HandlerFunc(PostHandler))
 	defer ts.Close()
 
 	bodyEcho := bytes.NewBufferString("Hello world!")
-
-	respPost, err := http.NewRequest(http.MethodPost, ts.URL, bodyEcho)
+	reqPost, err := http.NewRequest("GET", ts.URL, bodyEcho)
 	if err != nil {
 		t.Fatal(err)
 	}
-
-	respPost.Header.Set("Content-Type", "text/html")
 	client := &http.Client{}
-	req, err := client.Do(respPost)
+	req, err := client.Do(reqPost)
 	if err != nil {
 		t.Fatal(err)
 	}
-
 	outPost, err := io.ReadAll(req.Body)
 	if err != nil {
 		t.Fatal(err)
 	}
 
-	var errPost errResult
+	var errPost errPost
 	err = json.Unmarshal(outPost, &errPost)
 	if err != nil {
 		t.Fatal(err)
 	}
-	MsgErr := "Supported only GET method. Please use GET method."
+
+	MsgErr := "Supports only POST method. Please use POST method."
 	if errPost.ErrMsg != MsgErr {
 		t.Fatalf("Error should be: %s\n but i recieved: %s\n", MsgErr, errPost.ErrMsg)
 	}
