+++ handler/root.go
@@ -1,4 +1,4 @@
-package main
+package handler
 
 import (
 	"fmt"
@@ -6,7 +6,7 @@ import (
 	"time"
 )
 
-func rootHandler(w http.ResponseWriter, r *http.Request) {
+func RootHandler(w http.ResponseWriter, r *http.Request) {
 	t := time.Now()
 	fmt.Fprintln(w, "Hello World!\n", t.Format("02-01-2006 15:04:05"))
 }
