+++ handler/status.go
@@ -1,7 +1,6 @@
-package main
+package handler
 
 import (
-	"encoding/json"
 	"fmt"
 	"io"
 	"net/http"
@@ -14,7 +13,7 @@ type statusResponse struct {
 	ErrMsg string
 }
 
-func statusHandler(w http.ResponseWriter, r *http.Request) {
+func StatusHandler(w http.ResponseWriter, r *http.Request) {
 	vars := mux.Vars(r)
 	code := vars["codes"]
 
@@ -33,12 +32,4 @@ func statusHandler(w http.ResponseWriter, r *http.Request) {
 	}
 
 	w.WriteHeader(status)
-}
-func errMsg(msg string) string {
-	Msg := msgErr{msg}
-	outErr, err := json.Marshal(&Msg)
-	if err != nil {
-		return `{"Error code": "-1"}`
-	}
-	return string(outErr)
 }
